requirements.txt
Flask
Flask-CORS
PyMuPDF
python-docx
python-pptx
youtube-transcript-api
BeautifulSoup4
requests
langchain
openai
chromadb
sentence-transformers
google-generativeai
transformers
torch
accelerate
libretranslate
gTTS
pydub
reportlab